+++ Pandora/Paradigm/Inventory/Ability.hs
@@ -2,4 +2,4 @@ module Pandora.Paradigm.Inventory.Some.Ability (module Exports) where
 
 import Pandora.Paradigm.Inventory.Ability.Modifiable as Exports
 import Pandora.Paradigm.Inventory.Ability.Replaceable as Exports
-import Pandora.Paradigm.Inventory.Ability.Viewable as Exports
+import Pandora.Paradigm.Inventory.Ability.Gettable as Exports