+++ Pandora/Paradigm/Inventory/Ability/Modifiable.hs
@@ -1,9 +1,9 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 module Pandora.Paradigm.Inventory.Ability.Modifiable where
 
-import Pandora.Paradigm.Inventory.Ability.Viewable (Viewable)
+import Pandora.Paradigm.Inventory.Ability.Gettable (Gettable)
 import Pandora.Paradigm.Inventory.Ability.Replaceable (Replaceable)
 
-class (Viewable i, Replaceable i) => Modifiable i where
+class (Gettable i, Replaceable i) => Modifiable i where
 	type Modification i e r :: *
 	modify :: Modification i e r