+++ Pandora/Paradigm/Inventory/Some/Optics.hs
@@ -13,7 +13,7 @@ import Pandora.Pattern.Functor.Representable (Representable (Representation, (<#
 import Pandora.Pattern.Object.Setoid (Setoid ((==)))
 import Pandora.Paradigm.Controlflow.Effect.Conditional (Conditional ((?)))
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (run, (!)))
-import Pandora.Paradigm.Inventory.Ability.Viewable (Viewable (Viewing, view))
+import Pandora.Paradigm.Inventory.Ability.Gettable (Gettable (Getting, get))
 import Pandora.Paradigm.Inventory.Ability.Replaceable (Replaceable (Replacement, replace))
 import Pandora.Paradigm.Inventory.Ability.Modifiable (Modifiable (Modification, modify))
 import Pandora.Paradigm.Primary.Algebraic.Product ((:*:) ((:*:)))
@@ -109,18 +109,18 @@ instance Lensic Identity Maybe where
 			(Just target :*: mtb) -> Store ! Just target :*: ibs . Identity . mtb
 			(Nothing :*: _) -> Store ! Nothing :*: \_ -> source
 
-instance Viewable (Lens Identity) where
-	type instance Viewing (Lens Identity) source target = Lens Identity source target -> source -> target
-	view lens = extract @Identity . position @_ @(Store _) . run lens
+instance Gettable (Lens Identity) where
+	type instance Getting (Lens Identity) source target = Lens Identity source target -> source -> target
+	get lens = extract @Identity . position @_ @(Store _) . run lens
 
-instance Viewable (Lens Maybe) where
-	type instance Viewing (Lens Maybe) source target = Lens Maybe source target -> source -> Maybe target
-	view lens = position @_ @(Store _) . run lens
+instance Gettable (Lens Maybe) where
+	type instance Getting (Lens Maybe) source target = Lens Maybe source target -> source -> Maybe target
+	get lens = position @_ @(Store _) . run lens
 
 instance Pointable t => Replaceable (Lens t) where
 	type instance Replacement (Lens t) source target = target -> Lens t source target -> source -> source
 	replace new lens source = look @(t _) # point new # run lens source
 
-instance (Viewable (Lens t), Covariant (->) (->) t, Pointable t) => Modifiable (Lens t) where
+instance (Gettable (Lens t), Covariant (->) (->) t, Pointable t) => Modifiable (Lens t) where
 	type instance Modification (Lens t) source target = (target -> target) -> Lens t source target -> source -> source
 	modify f lens = extract . retrofit (f <-|-) . run lens