+++ Pandora/Paradigm/Structure/Ability/Zipper.hs
@@ -28,7 +28,7 @@ import Pandora.Paradigm.Schemes.P_Q_T (P_Q_T (P_Q_T))
 import Pandora.Paradigm.Structure.Ability.Morphable (Morphable, Morph (Rotate), Vertical (Up, Down))
 import Pandora.Paradigm.Structure.Ability.Substructure (Substructure (Available, Substance, substructure), Segment (Root), sub)
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (run, (!), (||=))
-import Pandora.Paradigm.Inventory.Ability.Viewable (view)
+import Pandora.Paradigm.Inventory.Ability.Gettable (get)
 import Pandora.Paradigm.Inventory.Ability.Replaceable (replace)
 import Pandora.Paradigm.Inventory.Some.Store (Store (Store))
 import Pandora.Paradigm.Inventory.Some.Optics (Lens, Convex)
@@ -92,7 +92,7 @@ instance (Covariant (->) (->) t, Semimonoidal (-->) (:*:) (:*:) t) => Substructu
 	type Available Left (Tape t <::> Tape t) = Identity
 	type Substance Left (Tape t <::> Tape t) = Tape t <::> Reverse t
 	substructure = P_Q_T ! \ii ->
-		let target = (view @(Convex Lens) (sub @Left) <-|-) ||= (lower ii) in
+		let target = (get @(Convex Lens) (sub @Left) <-|-) ||= (lower ii) in
 		let updated new = (replace @(Convex Lens) % sub @Left) <-|- new <-*- run (lower ii) in
 		Store ! Identity target :*: lift . (updated ||=) . extract
 
@@ -100,6 +100,6 @@ instance (Covariant (->) (->) t, Semimonoidal (-->) (:*:) (:*:) t) => Substructu
 	type Available Right (Tape t <::> Tape t) = Identity
 	type Substance Right (Tape t <::> Tape t) = Tape t <::> t
 	substructure = P_Q_T ! \ii ->
-		let target = (view @(Convex Lens) (sub @Right) <-|-) ||= lower ii in
+		let target = (get @(Convex Lens) (sub @Right) <-|-) ||= lower ii in
 		let updated new = (replace @(Convex Lens) % sub @Right) <-|- new <-*- run (lower ii) in
 		Store ! Identity target :*: lift . (updated ||=) . extract